move page settings up
