+++ server/src/database/database.module.ts
@@ -3,7 +3,7 @@ import { ConfigService } from '@nestjs/config';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { UserToken } from 'src/auth/user-tokens/user-token.entity';
 import { File as FileEntity } from 'src/files/file.entity';
-import { PageSetting } from 'src/pages/page-settings/page-setting.entity';
+import { PageSetting } from 'src/page-settings/page-setting.entity';
 import { Page } from 'src/pages/page.entity';
 import { Tier } from 'src/pages/tier.entity';
 import { Payment } from 'src/payments/payment.entity';
