+++ server/src/page-settings/page-settings.service.ts
@@ -1,0 +1,17 @@
+import { BadRequestException, Injectable } from '@nestjs/common';
+import { InjectRepository } from '@nestjs/typeorm';
+import { PageSetting } from './page-setting.entity';
+import { Repository } from 'typeorm';
+import { PagesService } from '../pages/pages.service';
+
+@Injectable()
+export class PageSettingsService {
+  constructor(
+    @InjectRepository(PageSetting) private repo: Repository<PageSetting>,
+    private pagesService: PagesService,
+  ) {}
+
+  update(pageId: number) {
+    if (!pageId) throw new BadRequestException('Page');
+  }
+}
