--- server/src/pages/page-settings/page-settings.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { PageSettingsService } from './page-settings.service';
-import { PageSettingsController } from './page-settings.controller';
-
-@Module({
-  providers: [PageSettingsService],
-  controllers: [PageSettingsController]
-})
-export class PageSettingsModule {}
