+++ server/src/pages/page.entity.ts
@@ -14,7 +14,7 @@ import {
   Unique,
 } from 'typeorm';
 import { Tier } from './tier.entity';
-import { PageSetting } from './page-settings/page-setting.entity';
+import { PageSetting } from '../page-settings/page-setting.entity';
 
 @Entity({ name: 'pages' })
 @Unique(['path'])
